Saner cursor handling on altscreen swapping
Normal and alternate screen both have 1-1 slots to save a cursor
position.

Old behaviour was putting the normal position in both save slots
on entering altscreen, which is probably not what was intended.

New behaviour is to only save normal position to normal slot.
Altscreen cursor is not saved on swapping, wouldn't make much
sense because altscreen is cleared before reentering.

Altscreen is changed to be cleared when entering the mode instead of
exiting.

Trying to react reasonably when the enter-exit-enter-exit cycle is
broken:
altscreen -> altscreen: clear the screen
  giving a fresh new screen if the last program forgot to exit
normal -> normal: doesn't do anything
  you can use '\e[1049l' in your shell prompt to not be stuck on
  the altscreen