Better font fallback
Sort once, don't substitute the pattern again and use match only for
the missing character from the sorted set. Before it matched against
the whole pattern, but this is not the best way as a default lang object
is added for the pattern in FcConfigSubstitute. FcFontSort handles it
intelligently: it ignores the languages already satisfied by previous
fonts. But running a match again with the whole pattern ignores the
already established order.